+++ src/gems/expression/print.py
@@ -17,12 +17,14 @@ from gems.expression.expression import (
     AbsNode,
     AllTimeSumNode,
     CeilNode,
+    DualNode,
     ExpressionNode,
     FloorNode,
     MaxNode,
     MinNode,
     PortFieldAggregatorNode,
     PortFieldNode,
+    ReducedCostNode,
     RoundNode,
     TimeEvalNode,
     TimeShiftNode,
@@ -135,6 +137,12 @@ class PrinterVisitor(ExpressionVisitor[str]):
 
     def minimum(self, node: MinNode) -> str:
         return "min(" + ", ".join(visit(op, self) for op in node.operands) + ")"
+
+    def dual(self, node: DualNode) -> str:
+        return f"dual({node.constraint_id})"
+
+    def reduced_cost(self, node: ReducedCostNode) -> str:
+        return f"reduced_cost({node.variable_id})"
 
 
 def print_expr(expression: ExpressionNode) -> str: